requirements.txt
pandas
numpy
requests
python-dotenv
geopandas
scikit-learn
joblib
pytest
matplotlib
seaborn
xgboost
lightgbm
